karma: declare stopEvilPortalDns, and stop the tier durations overflowing
The previous commit called stopEvilPortalDns() without declaring it: the
edit that was meant to add it to evil_portal.h anchored on a line that lives
in ble_common.h, so it silently changed nothing and the build failed.

That build also surfaced a real bug of the same shape as the baseDuration
one:

    warning: unsigned conversion from 'int' to 'uint16_t' changes value
    from '180000' to '48928'
      attackConfig.highTierDuration = 180000;

highTierDuration holds milliseconds in a uint16_t, so the high tier -- the
one reserved for the most promising targets -- ran for 49 seconds instead of
three minutes. The tier durations, PendingPortal::duration that they are
copied into, and getPortalDuration() that returns them are all 32 bit now.

+++ src/modules/wifi/karma_attack.cpp
@@ -1169,9 +1169,9 @@ AttackTier determineAttackTier(uint8_t priority) {
     return TIER_NONE;
 }
 
-uint16_t getPortalDuration(AttackTier tier) {
+uint32_t getPortalDuration(AttackTier tier) {
     switch (tier) {
-        case TIER_CLONE: return (uint16_t)attackConfig.cloneDuration;
+        case TIER_CLONE: return attackConfig.cloneDuration;
         case TIER_HIGH: return attackConfig.highTierDuration;
         case TIER_MEDIUM: return attackConfig.mediumTierDuration;
         case TIER_FAST: return attackConfig.fastTierDuration;
